one talker used
